Resolved merge conflicts

+++ django_project/search/static/js/search.js
@@ -567,8 +567,6 @@ APP.ResultGridView = Backbone.View.extend({
         var cur_pag_el = this.$el.find('#resultsPosition');
         var page_info = this.collection.pageInfo();
         var paginator = this._createSelectPaginator(page_info.current_page, page_info.pages);
-        var showing = page_info.limit;
-        if (page_info.total < page_info.limit) showing = page_info.total;
         var text = 'Page <span id="paginator"></span> of ' + page_info.pages + ' ('+page_info.total+' records)';
         if (page_info.current_page > 1) {
             $('#searchPrev').show();
@@ -584,7 +582,7 @@ APP.ResultGridView = Backbone.View.extend({
         cur_pag_el.html(text);
         $('#paginator').html(paginator);
         if (page_info.current_page == 1) {
-            $('#main-content').prepend('<div class="alert alert-info alert-dismissable alert-search"><button type="button" class="close" data-dismiss="alert" aria-hidden="true">&times;</button><strong>'+page_info.total+' Records Returned. Displaying '+showing+'.</strong></div>');
+            $('#main-content').prepend('<div class="alert alert-info alert-dismissable alert-search"><button type="button" class="close" data-dismiss="alert" aria-hidden="true">&times;</button><strong>'+page_info.total+' Records Returned. Displaying '+page_info.limit+'.</strong></div>');
             window.setTimeout(function () {
                 $(".alert").alert('close');
             }, 6000);
@@ -601,8 +599,7 @@ APP.ResultGridViewItem = Backbone.View.extend({
         'click': 'highlightResultItem',
         'click img': 'imagePopover',
         'mouseenter': 'focusItem',
-        'mouseleave': 'blurItem',
-        'mouseleave span': function() {console.log('test');},
+        'mouseleave': 'blurItem'
     },
     initialize: function() {
         $APP.on('highlightResultItem', $.proxy(this.highlightResultItem, this));
@@ -647,11 +644,11 @@ APP.ResultGridViewItem = Backbone.View.extend({
 
     blurItem: function() {
         if (typeof this.line != 'undefined') this.line.remove();
-        var selectedID = this.model.get('original_product_id');
+        var selectedID = this.model.get('id');
         if (APP.selectedFeatureID == selectedID) {
             $APP.trigger('highlightSearchRecord', {'original_product_id': selectedID, 'zoom': false});
         } else {
-            $APP.trigger('removeFocusFeature', {'original_product_id': selectedID});
+            $APP.trigger('removeFocusFeature', {'id': selectedID});
         }
     },
 
@@ -667,7 +664,7 @@ APP.ResultGridViewItem = Backbone.View.extend({
 
     highlightResultItem: function(event, data) {
         // if id is not set presume user has clicked in result panel on item
-        // if id is set presume user has clicked record on the map
+        // if id is set presuem user has clicked record on the map
         if (typeof data == 'undefined') {
             $APP.trigger('highlightSearchRecord', {'original_product_id': this.model.get('original_product_id'), 'zoom': true});
             var selectedID = this.model.get('original_product_id');
@@ -744,9 +741,9 @@ var template = [
               '<% } else { %>UNK',
               '<% } %>',
             '</p></div>',
-            '<span class="button metadata-button btn btn-default" data-title="View Metadata"><i class="icon-list-alt"></i></span>',
-            '<span class="button cart-button btn btn-default" data-title="Add to Cart"><i class="icon-shopping-cart"></i></span>',
-            '<span class="button cart-remove-button btn btn-danger hide" data-title="Remove From Cart"><i class="icon-remove"></i></span>',
+            '<span class="button metadata-button btn btn-default" data-toggle="tooltip" data-title="View Metadata"><i class="icon-list-alt"></i></span>',
+            '<span class="button cart-button btn btn-default" data-toggle="tooltip" data-title="Add to Cart"><i class="icon-shopping-cart"></i></span>',
+            '<span class="button cart-remove-button btn btn-danger hide" data-toggle="tooltip" data-title="Remove From Cart"><i class="icon-remove"></i></span>',
           '</div>'
           ].join('');
 
@@ -793,10 +790,13 @@ APP.CartGridView = Backbone.View.extend({
     },
 
     deleteItem: function(event, data) {
-        var exist = APP.Cart.find(function(item) {
-            return item.get("product").original_product_id == data.id;
-        });
-        exist.destroy({wait: true});
+        var del = confirm('Are you sure?');
+        if (del) {
+            var exist = APP.Cart.find(function(item) {
+                return item.get("product").original_product_id == data.id;
+            });
+            exist.destroy({wait: true});
+        }
     },
 
     render: function() {
@@ -835,8 +835,11 @@ APP.CartGridViewItem = Backbone.View.extend({
         });
     },
     delete: function() {
-        $APP.trigger('removedItemFromCartUpdateResults', {'original_product_id': this.model.get('product').original_product_id});
-        this.model.destroy({wait: true});
+        var del = confirm('Are you sure?');
+        if (del) {
+            $APP.trigger('removedItemFromCartUpdateResults', {'original_product_id': this.model.get('product').original_product_id});
+            this.model.destroy({wait: true});
+        }
     },
     render: function() {
        $(this.el).html(_.template(templateCart, {model:this.model}));
